+++ crates/canisters/blank/src/lib.rs
@@ -2,7 +2,7 @@
 
 use icu::{
     Error,
-    canister::EXAMPLE,
+    canister::BLANK,
     ops::{request::create_canister_request, response::CreateCanisterResponse},
     prelude::*,
 };
@@ -11,7 +11,7 @@ use icu::{
 // ICU
 //
 
-icu_start!(EXAMPLE);
+icu_start!(BLANK);
 
 const fn icu_setup() {}
 async fn icu_install(_: Option<Vec<u8>>) {}
@@ -20,7 +20,7 @@ async fn icu_upgrade() {}
 // create_example
 #[update]
 async fn create_example() -> Result<CreateCanisterResponse, Error> {
-    create_canister_request::<()>(&EXAMPLE, None).await
+    create_canister_request::<()>(&BLANK, None).await
 }
 
 export_candid!();