+++ crates/canisters/delegation/src/lib.rs
@@ -1,21 +1,15 @@
 #![allow(clippy::unused_async)]
 
-use icu::{canister::GAME, prelude::*};
+use icu::{canister::DELEGATION, prelude::*};
 
 //
 // ICU
 //
 
-icu_start!(GAME);
+icu_start!(DELEGATION);
 
 const fn icu_setup() {}
 async fn icu_install(_: Option<Vec<u8>>) {}
 async fn icu_upgrade() {}
-
-// Minimal endpoint for visibility
-#[query]
-const fn name() -> &'static str {
-    "icu:game"
-}
 
 export_candid!();