+++ crates/canisters/root/src/lib.rs
@@ -27,32 +27,32 @@ async fn icu_upgrade() {}
 // WASMS
 pub static WASMS: &[(CanisterType, &[u8])] = &[
     (
-        canister::EXAMPLE,
+        canister::BLANK,
         #[cfg(icu_github_ci)]
         &[],
         #[cfg(not(icu_github_ci))]
-        include_bytes!("../../../../.dfx/local/canisters/example/example.wasm.gz"),
+        include_bytes!("../../../../.dfx/local/canisters/blank/blank.wasm.gz"),
     ),
     (
-        canister::PLAYER_HUB,
+        canister::DELEGATION,
         #[cfg(icu_github_ci)]
         &[],
         #[cfg(not(icu_github_ci))]
-        include_bytes!("../../../../.dfx/local/canisters/player_hub/player_hub.wasm.gz"),
+        include_bytes!("../../../../.dfx/local/canisters/delegation/delegation.wasm.gz"),
     ),
     (
-        canister::GAME,
+        canister::SHARDER,
         #[cfg(icu_github_ci)]
         &[],
         #[cfg(not(icu_github_ci))]
-        include_bytes!("../../../../.dfx/local/canisters/game/game.wasm.gz"),
+        include_bytes!("../../../../.dfx/local/canisters/sharder/sharder.wasm.gz"),
     ),
     (
-        canister::INSTANCE,
+        canister::SHARD,
         #[cfg(icu_github_ci)]
         &[],
         #[cfg(not(icu_github_ci))]
-        include_bytes!("../../../../.dfx/local/canisters/instance/instance.wasm.gz"),
+        include_bytes!("../../../../.dfx/local/canisters/shard/shard.wasm.gz"),
     ),
 ];
 
@@ -60,15 +60,10 @@ pub static WASMS: &[(CanisterType, &[u8])] = &[
 /// ENDPOINTS
 ///
 
-// create_example (demo)
+// create_blank
 #[update]
-async fn create_example() -> Result<CreateCanisterResponse, Error> {
-    create_canister_request::<()>(&canister::EXAMPLE, None).await
-}
-
-#[update]
-async fn get_icp_rate() -> Result<f64, Error> {
-    icu::interface::ic::get_icp_xdr_conversion_rate().await
+async fn create_blank() -> Result<CreateCanisterResponse, Error> {
+    create_canister_request::<()>(&canister::BLANK, None).await
 }
 
 // end