+++ crates/canisters/shard/src/lib.rs
@@ -1,0 +1,15 @@
+#![allow(clippy::unused_async)]
+
+use icu::{canister::SHARD, prelude::*};
+
+//
+// ICU
+//
+
+icu_start!(SHARD);
+
+const fn icu_setup() {}
+async fn icu_install(_: Option<Vec<u8>>) {}
+async fn icu_upgrade() {}
+
+export_candid!();