+++ crates/icu/src/canister.rs
@@ -1,6 +1,6 @@
 use crate::types::CanisterType;
 
-pub const EXAMPLE: CanisterType = CanisterType::new("example");
-pub const GAME: CanisterType = CanisterType::new("game");
-pub const INSTANCE: CanisterType = CanisterType::new("instance");
-pub const PLAYER_HUB: CanisterType = CanisterType::new("player_hub");
+pub const BLANK: CanisterType = CanisterType::new("blank");
+pub const DELEGATION: CanisterType = CanisterType::new("delegation");
+pub const SHARDER: CanisterType = CanisterType::new("sharder");
+pub const SHARD: CanisterType = CanisterType::new("shard");